src/sample-chat-app/api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
azure-ai-agents==1.1.0
azure-ai-projects==1.0.0
azure-core==1.36.0
azure-identity==1.25.1
azure-storage-blob==12.27.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
openai==2.9.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
